Fix 32/64-bit .DLL issue
Problem was WinDLL was trying to open a 32-bit .DLL on a 64-bit machine.
Fix is to
1) Assume windows
2) Consolidate all calls to WinDLL()
3) Use sys.maxsize to determine if the current machine
is 32 or 64 bits and open the corresponding .DLL.
Add helpful instructions if the .DLL load still fails.

Add a friendly message about the assumption of Windows.

Add Windows OS classifier in the package manifest.